pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fabric-ci-test"
version = "0.1.0"
description = "Power BI visual regression testing CLI for Microsoft Fabric CI/CD"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [{name = "Filippos A. Dorezi", email = "filippos.dorezi@gmail.com"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "playwright>=1.40",
    "pytest>=7.0",
    "pytest-html>=4.0",
    "pytest-playwright>=0.7",
    "pytest-xdist>=3.0",
    "python-dotenv>=1.0",
    "requests>=2.28",
]

[project.urls]
Homepage = "https://github.com/FilippDor/Fabric-UI-testing"
Repository = "https://github.com/FilippDor/Fabric-UI-testing"
Issues = "https://github.com/FilippDor/Fabric-UI-testing/issues"

[tool.setuptools.packages.find]
include = ["fabric_ci_test*", "helper_functions*"]
